Move some debug endpoints from repo-updater to frontend
They're not specific to repo-updater internals and can live anywhere. This moves them away from it, in an attempt to make the service leaner.
