+++ cmd/repo-updater/shared/debug.go
@@ -19,24 +19,6 @@ func createDebugServerEndpoints(ready chan struct{}, debugserverEndpoints *LazyD
 			}),
 		},
 		{
-			Name: "List Authz Providers",
-			Path: "/list-authz-providers",
-			Handler: http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-				// wait until we're healthy to respond
-				<-ready
-				// listAuthzProvidersEndpoint is guaranteed to be assigned now
-				debugserverEndpoints.listAuthzProvidersEndpoint(w, r)
-			}),
-		},
-		{
-			Name: "Gitserver Repo Status",
-			Path: "/gitserver-repo-status",
-			Handler: http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-				<-ready
-				debugserverEndpoints.gitserverReposStatusEndpoint(w, r)
-			}),
-		},
-		{
 			Name: "Manual Repo Purge",
 			Path: "/manual-purge",
 			Handler: http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
